Move to a separate test file gpuarray tests that must always run, even when pygpu isn't available
