+++ theano/sandbox/gpuarray/tests/test_type.py
@@ -1,6 +1,5 @@
 import os
 
-from nose.tools import assert_raises
 import numpy
 
 import theano
@@ -11,12 +10,9 @@ from theano.misc.pkl_utils import CompatUnpickler
 
 from .config import test_ctx_name
 from .test_basic_ops import rand_gpuarray
-
 from ..type import GpuArrayType
-from .. import pygpu_activated
 
-if pygpu_activated:
-    import pygpu
+import pygpu
 
 
 def test_deep_copy():
@@ -56,6 +52,7 @@ def test_filter_float():
 
 
 def test_unpickle_gpuarray_as_numpy_ndarray_flag0():
+    """ Test when pygpu isn't there for unpickle are in test_pickle.py"""
     oldflag = config.experimental.unpickle_gpu_on_cpu
     config.experimental.unpickle_gpu_on_cpu = False
 
@@ -68,33 +65,8 @@ def test_unpickle_gpuarray_as_numpy_ndarray_flag0():
                 u = CompatUnpickler(fp, encoding="latin1")
             else:
                 u = CompatUnpickler(fp)
-            if pygpu_activated:
-                mat = u.load()
-                assert isinstance(mat, pygpu.gpuarray.GpuArray)
-                assert numpy.asarray(mat)[0] == -42.0
-            else:
-                assert_raises(ImportError, u.load)
-    finally:
-        config.experimental.unpickle_gpu_on_cpu = oldflag
-
-
-def test_unpickle_gpuarray_as_numpy_ndarray_flag1():
-    oldflag = config.experimental.unpickle_gpu_on_cpu
-    config.experimental.unpickle_gpu_on_cpu = True
-
-    try:
-        testfile_dir = os.path.dirname(os.path.realpath(__file__))
-        fname = 'GpuArray.pkl'
-
-        with open(os.path.join(testfile_dir, fname), 'rb') as fp:
-            if PY3:
-                u = CompatUnpickler(fp, encoding="latin1")
-            else:
-                u = CompatUnpickler(fp)
             mat = u.load()
-
-        assert isinstance(mat, numpy.ndarray)
-        assert mat[0] == -42.0
-
+            assert isinstance(mat, pygpu.gpuarray.GpuArray)
+            assert numpy.asarray(mat)[0] == -42.0
     finally:
         config.experimental.unpickle_gpu_on_cpu = oldflag
